fixed loader

+++ src/main/java/com/se2/kursplaner/event/DateiLoader.java
@@ -60,7 +60,7 @@ public class DateiLoader implements CommandLineRunner{
                 String studiengangName = (String) studiengang.get("StudingangName");
                 Studiengang studiengangData = studiengangRepository.findByName(studiengangName);
                 String studiengangKuerzel = (String) studiengang.get("StudiengangKuerzel");
-                int studiengangSemesteranzahl = (int) studiengang.get("StudiengangSemesteranzahl");
+                Integer studiengangSemesteranzahl = ((Long) studiengang.get("StudiengangSemesteranzahl")).intValue();
 
                 //Studingang erstellen wenn nicht vorhanden
                 if(studiengangData == null){
